# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CaseStatus = (typeof CaseStatus)[keyof typeof CaseStatus];

export const CaseStatus = {
  open: "open",
  closed: "closed",
  archived: "archived",
  active: "active",
} as const;

export interface Case {
  id: number;
  title: string;
  description?: string;
  status: CaseStatus;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export type CreateCaseInputStatus =
  (typeof CreateCaseInputStatus)[keyof typeof CreateCaseInputStatus];

export const CreateCaseInputStatus = {
  open: "open",
  closed: "closed",
  archived: "archived",
  active: "active",
} as const;

export interface CreateCaseInput {
  title: string;
  description?: string;
  status?: CreateCaseInputStatus;
  tags?: string[];
}

export type UpdateCaseInputStatus =
  (typeof UpdateCaseInputStatus)[keyof typeof UpdateCaseInputStatus];

export const UpdateCaseInputStatus = {
  open: "open",
  closed: "closed",
  archived: "archived",
  active: "active",
} as const;

export interface UpdateCaseInput {
  title?: string;
  description?: string;
  status?: UpdateCaseInputStatus;
  tags?: string[];
}

export type EntityType = (typeof EntityType)[keyof typeof EntityType];

export const EntityType = {
  person: "person",
  organization: "organization",
  company: "company",
  nonprofit: "nonprofit",
  government_agency: "government_agency",
  program: "program",
  facility: "facility",
  location: "location",
  event: "event",
} as const;

export interface Entity {
  id: number;
  name: string;
  type: EntityType;
  description?: string;
  aliases?: string[];
  caseId?: number;
  createdAt: string;
}

export interface Document {
  id: number;
  title: string;
  filePath?: string;
  source?: string;
  publishDate?: string;
  uploadedAt: string;
  caseId?: number;
}

export interface TimelineEntry {
  id: number;
  title: string;
  description?: string;
  eventDate: string;
  linkedEntityId?: number;
  linkedEntityName?: string;
  linkedDocumentId?: number;
  caseId?: number;
  createdAt: string;
}

export interface Event {
  id: number;
  title: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  source?: string;
  timestamp: string;
  caseId?: number;
  createdAt: string;
}

export interface Note {
  id: number;
  caseId: number;
  content: string;
  createdAt: string;
  updatedAt?: string;
}

export interface Relationship {
  id: number;
  entityAId: number;
  entityBId: number;
  entityAName?: string;
  entityBName?: string;
  relationshipType: string;
  evidenceDocumentId?: number;
  confidence?: number;
  caseId?: number;
  createdAt: string;
}

export interface MoneyFlow {
  id: number;
  sourceEntityId: number;
  destinationEntityId: number;
  sourceEntityName?: string;
  destinationEntityName?: string;
  amount: number;
  date?: string;
  description?: string;
  supportingDocumentId?: number;
  caseId?: number;
  createdAt: string;
}

export interface CaseSummary {
  case: Case;
  entities: Entity[];
  documents: Document[];
  timeline: TimelineEntry[];
  events: Event[];
  notes: Note[];
  relationships: Relationship[];
  moneyFlows: MoneyFlow[];
}

export type CreateEntityInputType =
  (typeof CreateEntityInputType)[keyof typeof CreateEntityInputType];

export const CreateEntityInputType = {
  person: "person",
  organization: "organization",
  company: "company",
  nonprofit: "nonprofit",
  government_agency: "government_agency",
  program: "program",
  facility: "facility",
  location: "location",
  event: "event",
} as const;

export interface CreateEntityInput {
  name: string;
  type: CreateEntityInputType;
  description?: string;
  aliases?: string[];
  caseId?: number;
}

export type UpdateEntityInputType =
  (typeof UpdateEntityInputType)[keyof typeof UpdateEntityInputType];

export const UpdateEntityInputType = {
  person: "person",
  organization: "organization",
  company: "company",
  nonprofit: "nonprofit",
  government_agency: "government_agency",
  program: "program",
  facility: "facility",
  location: "location",
  event: "event",
} as const;

export interface UpdateEntityInput {
  name?: string;
  type?: UpdateEntityInputType;
  description?: string;
  aliases?: string[];
}

export interface EntityProfile {
  entity: Entity;
  relationships: Relationship[];
  documents: Document[];
  timelineAppearances: TimelineEntry[];
}

export interface CreateDocumentInput {
  title: string;
  filePath?: string;
  source?: string;
  publishDate?: string;
  caseId?: number;
}

export interface CreateRelationshipInput {
  entityAId: number;
  entityBId: number;
  relationshipType: string;
  evidenceDocumentId?: number;
  confidence?: number;
  caseId?: number;
}

export interface CreateTimelineEntryInput {
  title: string;
  description?: string;
  eventDate: string;
  linkedEntityId?: number;
  linkedDocumentId?: number;
  caseId?: number;
}

export interface CreateEventInput {
  title: string;
  location?: string;
  latitude?: number;
  longitude?: number;
  source?: string;
  timestamp: string;
  caseId?: number;
}

export interface CreateNoteInput {
  caseId: number;
  content: string;
}

export interface UpdateNoteInput {
  content: string;
}

export interface CreateMoneyFlowInput {
  sourceEntityId: number;
  destinationEntityId: number;
  amount: number;
  date?: string;
  description?: string;
  supportingDocumentId?: number;
  caseId?: number;
}

export type ListEntitiesParams = {
  caseId?: number;
};

export type ListDocumentsParams = {
  caseId?: number;
};

export type UploadDocumentBody = {
  file?: Blob;
  caseId?: number;
  title?: string;
  source?: string;
  publishDate?: string;
};

export type ListRelationshipsParams = {
  entityId?: number;
  caseId?: number;
};

export type ListTimelineEntriesParams = {
  caseId?: number;
};

export type ListEventsParams = {
  caseId?: number;
};

export type ListNotesParams = {
  caseId: number;
};

export type ListMoneyFlowsParams = {
  caseId?: number;
};