# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS Investigative Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /relationships:
    get:
      operationId: listRelationships
      tags: [relationships]
      summary: List relationships
      parameters:
        - name: entityId
          in: query
          schema:
            type: integer
        - name: caseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Relationship"
    post:
      operationId: createRelationship
      tags: [relationships]
      summary: Create a relationship
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelationshipInput"
      responses:
        "201":
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Relationship"
  /relationships/{id}:
    delete:
      operationId: deleteRelationship
      tags: [relationships]
      summary: Delete a relationship
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /timeline:
    get:
      operationId: listTimelineEntries
      tags: [timeline]
      summary: List timeline entries
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Timeline entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEntry"
    post:
      operationId: createTimelineEntry
      tags: [timeline]
      summary: Create a timeline entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEntryInput"
      responses:
        "201":
          description: Entry created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEntry"
  /timeline/{id}:
    delete:
      operationId: deleteTimelineEntry
      tags: [timeline]
      summary: Delete a timeline entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes for a case
      parameters:
        - name: caseId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    put:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /money-flows:
    get:
      operationId: listMoneyFlows
      tags: [money-flows]
      summary: List money flows
      parameters:
        - name: caseId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Money flows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MoneyFlow"
    post:
      operationId: createMoneyFlow
      tags: [money-flows]
      summary: Create a money flow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoneyFlowInput"
      responses:
        "201":
          description: Money flow created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoneyFlow"
  /money-flows/{id}:
    delete:
      operationId: deleteMoneyFlow
      tags: [money-flows]
      summary: Delete a money flow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Relationship:
      type: object
      properties:
        id:
          type: integer
        entityAId:
          type: integer
        entityBId:
          type: integer
        entityAName:
          type: string
        entityBName:
          type: string
        relationshipType:
          type: string
        evidenceDocumentId:
          type: integer
        confidence:
          type: number
        caseId:
          type: integer
        createdAt:
          type: string
      required: [id, entityAId, entityBId, relationshipType, createdAt]
    CreateRelationshipInput:
      type: object
      properties:
        entityAId:
          type: integer
        entityBId:
          type: integer
        relationshipType:
          type: string
        evidenceDocumentId:
          type: integer
        confidence:
          type: number
        caseId:
          type: integer
      required: [entityAId, entityBId, relationshipType]
    TimelineEntry:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        linkedEntityId:
          type: integer
        linkedEntityName:
          type: string
        linkedDocumentId:
          type: integer
        caseId:
          type: integer
        createdAt:
          type: string
      required: [id, title, eventDate, createdAt]
    CreateTimelineEntryInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventDate:
          type: string
        linkedEntityId:
          type: integer
        linkedDocumentId:
          type: integer
        caseId:
          type: integer
      required: [title, eventDate]
...
    Note:
      type: object
      properties:
        id:
          type: integer
        caseId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, caseId, content, createdAt]
    CreateNoteInput:
      type: object
      properties:
        caseId:
          type: integer
        content:
          type: string
      required: [caseId, content]
    UpdateNoteInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
    MoneyFlow:
      type: object
      properties:
        id:
          type: integer
        sourceEntityId:
          type: integer
        destinationEntityId:
          type: integer
        sourceEntityName:
          type: string
        destinationEntityName:
          type: string
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        supportingDocumentId:
          type: integer
        caseId:
          type: integer
        createdAt:
          type: string
      required: [id, sourceEntityId, destinationEntityId, amount, createdAt]
    CreateMoneyFlowInput:
      type: object
      properties:
        sourceEntityId:
          type: integer
        destinationEntityId:
          type: integer
        amount:
          type: number
        date:
          type: string
        description:
          type: string
        supportingDocumentId:
          type: integer
        caseId:
          type: integer
      required: [sourceEntityId, destinationEntityId, amount]