# lib/api-zod/src/generated/types/updateNoteInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NEXUS Investigative Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateNoteInput {
  content: string;
}